# refactor: OpenCell.shadyCellsの実装をmapから[]Positionsにした