--- a/go.mod
+++ b/go.mod
@@ -3,3 +3,8 @@ module github.com/inahym196/bomb
 go 1.23.5
 
 require github.com/MakeNowJust/heredoc v1.0.0
+
+require (
+	github.com/aclements/go-moremath v0.0.0-20241023150245-c8bbc672ef66 // indirect
+	golang.org/x/perf v0.0.0-20250305200902-02a15fd477ba // indirect
+)