```bash
❯ benchstat before.out after.out                                                             
goos: darwin
goarch: amd64
pkg: github.com/inahym196/bomb/internal/solver/domain
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
                 │  before.out  │              after.out               │
                 │    sec/op    │    sec/op     vs base                │
NewSolverCells-8   451.6n ± 10%   344.1n ± 35%  -23.80% (p=0.009 n=10)

                 │ before.out │           after.out            │
                 │    B/op    │    B/op     vs base            │
NewSolverCells-8   720.0 ± 0%   720.0 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                 │ before.out │             after.out              │
                 │ allocs/op  │ allocs/op   vs base                │
NewSolverCells-8   6.000 ± 0%   4.000 ± 0%  -33.33% (p=0.000 n=10)
```